Update functional
The note naming system has been replaced. Now the note name != file name. The note name is stored in the metadata, just like the deadline.
The note ToDo file is combined with the metadata file, since both have the JSON extension and the note files and metadata files to them have the same base name.