Added functionality for notification about the upcoming deadline of a note.
True, for now the time before the start of the notification is fixed in the code, then it should be taken out to the settings or to the metadata of the note itself.Or both.
Depending on whether the time before the start of the notification is set when creating the note itself or not - use the default value - from the settings.