Automatic update of the model and view has been implemented.
After closing the creation of a note (specifically the widget container), the parent method is called, i.e. the class with the view, which first calls the update method of its parent, i.e. TrayMenu, and then receives the updated (reinitialized) model and passes it to the view again.

The settings have been added to set the time before notification of the note deadline.

You can set the time (in minutes) before the deadline expires to notify the user about it.

Added import/export of notes.

It is advisable to finish this system to avoid errors, as well as add protection in the form of writing/reading a certain structure (to avoid importing unnecessary information if the user has thrown something else into the directory).